feat(refactorExam): complete refactor of all ExamPages into reusable components. Rename files to remove section numbers.
