+++ app/components/ExamPages/ScalesNotateTemplate.tsx
@@ -63,6 +63,11 @@ export default function ScalesNotation({
         <Grid container spacing={4} p={2}>
           <Grid item xs={4}>
             <Stack gap={2} alignItems={"center"}>
+              {page === 6 && (
+                <Typography variant="h6" align="center">
+                  Section 3: Notate Scales
+                </Typography>
+              )}
               <Box
                 width={273}
                 height={456}
